fixed covariance warning
Code is included that will test if a matrix is positive definite. The covm matrices are all positive definite so the fact that the warning is occurring is simply for off diagonal elements > 0. Was apparently patched so don't know why that's happening here, however have included a tol = 1e-6 to stop the warning.